+++ tests/gen_ref_images.py
@@ -6,6 +6,7 @@
 
 # Standard library imports
 from __future__ import print_function
+import glob
 import itertools
 import math
 import os
@@ -21,8 +22,6 @@ import pplot
 ###
 # Relative figure size
 SCALE = 5
-# Number of reference image sets
-SETS = 12
 
 
 ###
@@ -36,10 +35,9 @@ def def_file_names(mode, ref_dir, test_dir, img_name, olist):
     else:
         # Support incomplete sets of reference images to reduce size
         # of source tar ball
-        print(ref_dir)
+        fdirs = glob.glob(ref_dir.format('*'))
         temp_ref_fname = [
-            os.path.realpath(os.path.join(ref_dir.format(num), img_name))
-            for num in range(1, SETS+1)
+            os.path.realpath(os.path.join(fdir, img_name)) for fdir in fdirs
         ]
         ref_fname = [item for item in temp_ref_fname if os.path.exists(item)]
     olist.append({'ref_fname':ref_fname, 'test_fname':test_fname})